Add new class for pipeline
