+++ src/main/java/ru/cs/eatright/nlpmodel/Chunker.java
@@ -5,6 +5,7 @@ import edu.stanford.nlp.ling.CoreLabel;
 import edu.stanford.nlp.ling.tokensregex.TokenSequenceMatcher;
 import edu.stanford.nlp.ling.tokensregex.TokenSequencePattern;
 import edu.stanford.nlp.util.CoreMap;
+import ru.cs.eatright.nlpmodel.signatures.Phrase;
 
 import java.util.ArrayList;
 import java.util.Collections;
