+++ src/main/java/ru/cs/eatright/nlpmodel/StopWordTokenizer.java
@@ -7,6 +7,7 @@ import org.apache.lucene.analysis.ru.RussianAnalyzer;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.cs.eatright.nlpmodel.signatures.Token;
 
 import java.io.IOException;
 import java.util.ArrayList;
