+++ src/main/java/ru/cs/eatright/nlpmodel/Tokenizer.java
@@ -1,5 +1,7 @@
 package ru.cs.eatright.nlpmodel;
 
+import ru.cs.eatright.nlpmodel.signatures.Token;
+
 import java.io.IOException;
 import java.util.List;
