+++ src/test/java/ru/cs/eatright/parsing/PipelineTest.java
@@ -1,8 +1,8 @@
-package ru.cs.eatright.nlp;
+package ru.cs.eatright.parsing;
 
 import org.junit.Test;
 import ru.cs.eatright.nlpmodel.*;
-import ru.cs.eatright.parsing.QueryPipeline;
+import ru.cs.eatright.nlpmodel.signatures.Query;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -14,37 +14,7 @@ import static org.junit.Assert.*;
 public class PipelineTest {
 
     private final Tokenizer tokenizer = new StopWordTokenizer();
-    private final Chunker chunker = new Chunker();
     private final QueryPipeline pipeline = new QueryPipeline();
-
-    /*private static void pipelineAssert(List<Phrase> expected, List<Phrase> actual) {
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void extractNounPhrases1() throws Exception {
-        List<Phrase> expected = Arrays.asList(
-                new Phrase(Collections.singletonList("вася"), Collections.singletonList("NN")),
-                new Phrase(Arrays.asList("вкусную", "кашу"), Arrays.asList("VB", "NN"))
-        );
-        List<Phrase> actual = chunker.getPhrases(tokenizer.tokenize("Вася ест вкусную кашу"));
-
-        pipelineAssert(expected, actual);
-    }
-
-    @Test
-    public void extractNounPhrases2() throws Exception {
-        List<Phrase> expected = Arrays.asList(
-                new Phrase(Arrays.asList("съем", "грибов"), Arrays.asList("NN", "NN")),
-                new Phrase(Arrays.asList("жареной", "картошкой"), Arrays.asList("NN", "NN")),
-                new Phrase(Collections.singletonList("курицей"), Collections.singletonList("NN"))
-        );
-
-        List<Phrase> actual = pipeline.extractNounPhrases(
-                pipeline.tokenizeAndCleanText("Съем грибов с жареной картошкой и курицей", false));
-
-        pipelineAssert(expected, actual);
-    }*/
 
     @Test
     public void stemQueries1() throws IOException {
